ocr-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.1.0
numpy==2.2.1
opencv-python==4.10.0.84
easyocr==1.7.2